requirements.txt
# Core dependencies
python-binance>=1.0.16
python-telegram-bot>=20.0
python-dotenv>=0.19.0

# Data processing and analysis
pandas>=1.3.0
numpy>=1.20.0